requirements.txt
altair>=4.2.0
matplotlib>=3.5.1
numpy>=1.21.5
pandas>=1.3.5
PyYAML>=5.3.1
streamlit>=1.7.0
